refactor: extract agent loading helpers
